Created the Atbash method
The Atbash method converts the PlainText string into a character array, then loops through every letter.
In each iteration of the loop, the current letter is encoded using the ReverseLetter method.
The new letter is the added to the CipherTextLetters char array. This array is later converted to a string then returned.

- Created the Atbash method within the AtbashCipher class

Signed-off-by: = <81330314+lightman210567@users.noreply.github.com>