fixed an error in the Atbash method
The Atbash was returning IndexOutOfRange exceptions on every run.
This was due to the i++ being placed before the letter is added to the array - this causes the final letter to fail.
This issue is corrected in this commit by moving the i++ to after where the letter is added to the cipher text array.

- Fixed an error in the Atbash method

modified: Encipher/Encipher.sln
modified: Encipher/Encipher/AtbashCipher.cs

Signed-off-by: = <81330314+lightman210567@users.noreply.github.com>