fixed the atbash method string return
The conversion of the CipherTextLetters[] array to string was causing the unit tests to fail.
This was because System.Char[] was being returned instead of the cipher text.
This was fixed by changing the conversion code.

The unit tests were also broken, as the Assert.AreEqual() had its imputs the wrong way round.
The inputs were swapped around and now the tests function correctly.

- Fixed the conversion of atbash letters to string
- Fixed the AreAqual assert in the atbash unit tests

Signed-off-by: = <81330314+lightman210567@users.noreply.github.com>